Add additional cleanup tasks and checks
Legacy environments, depending on age, may have services installed on
hosts. If this is the case, cleanup will miss host machines because we
assume services are installed in containers. This change adds additional
cleanup tasks to ensure we get everything.

This change also introduces overlay inventories to our process. This
allows us to deploy services with minimal, if any, change to a running
environment.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>