Vendor an overlay inventory and remove elk_all meta groups
The overlay inventory we use is maintained by upstream. This change vendors the
overlay inventory which ensures our version doesn't change out from under us.
This is a precautionary measure being taken to ensure consistent deployments.

This change also removes the "elk_all" meta-group from the container cleanup
process which covers beat installations from the elastic-stack. The inclusion
of this meta-group could result in the unnecessary destruction of other
services. The only groups now included in the container cleanup process are
groups that should have no relevance a typical deployment as it pertains to our
environmental tools.

Signed-off-by: Kevin Carter <kevin@cloudnull.com>