Add support for authenticated proxies
This change will now parse the proxy URL for both http and https. If the
proxy url has an "@" sign within it it will enable authenticated proxy
support. This has been setup such that it will check for authenticated
proxies for both http and https independently as a deployment may have
an authenicated proxy for specific protocols.

Signed-off-by: cloudnull <kevin@cloudnull.com>